standalone: updated test case and fixed bugs in gem5.

+++ src/apps/standalone/hw/platform.cc
@@ -1,3 +1,4 @@
+#include "common.h"
 #include "platform.h"
 
 //testcase specific defines
@@ -7,7 +8,7 @@
 #define TESTCASE_RESULT_ADDR 0x00041000
 
 volatile uint64_t *ui64_ptr = (uint64_t*)TESTCASE_RESULT_ADDR;
-volatile uintptr_t *ui32_ptr = (uintptr_t*)TESTCASE_RESULT_ADDR;
+volatile uint32_t *ui32_ptr = (uintptr_t*)TESTCASE_RESULT_ADDR;
 
 void init() {
     ui64_ptr[0] = TESTCASE_FAILED;
